Fix: parse functions with positional args in exp.func
We previously used to construct functions with positional arguments by
passing them directly to the corresponding from_arg_list method.

This was problematic because for some functions the argument order varies
between dialects (e.g. LOG) and since we weren't actually parsing the
function but instead constructing it directly, the produced AST was wrong.

This fixes the issue by producing a string that represents the function call
in the target dialect and then parsing it in that dialect to get the right AST.