Feat!: introduce Dialect settings, make MySQL case-sensitive by default
This introduces a way for SQLGlot users to specify Dialect-level settings,
by including them in the dialect string: "dialect, key1 = value1, ...".

Now, Dialect.get_or_raise will return a Dialect instance by default, instead
of a type, to ensure that the user-supplied settings override the defaults.

The motivation behind this is that we needed to provide a way to override the
normalization strategy of MySQL, because we realized that it depends on the
underlying OS and hence a single default behavior cannot capture all possible
strategies employed by different systems.

This feature creates a way to add more dialect-specific settings in the future,
such as the dialect's version etc.

Finally, MySQL's default normalization strategy has been changed so that
all identifiers are treated as case-sensitive, regardless of being quoted.
This matches MySQL's behavior on Linux systems, so we decided to use that
as a default. Users that require a different strategy can simply specify:

    dialect = "mysql, normalization_strategy = lowercase"

This will match MySQL's behavior in MacOS and Windows systems, for example.