Fix!(bigquery): transpile brackets to dot notation less aggressively (#3084)
We used to transpile doc["field"] into doc.field in BQ, because it
doesn't support the former syntax for struct values. However, it
does support it for JSON values, and so that conversion would break
valid queries. This only applies said transpilation if we're sure
that the left-hand side of the bracket is a STRUCT value.