Refactor!(optimizer): don't create scopes for DDLs
We used to create Scopes for both Query and DDL expressions, but the
logic that handled the latter was brittle, incomplete and resulted in
another expression type that users of the Scope module had to take
into account, thus making the interface more complex to use.

This makes it so that we'll only optimize the _queries_ of statements
such as CTAS and INSERT INTO, instead of the whole thing. Same holds
for lineage inference - we'll only analyze the DDL queries.