Updated tokens.py
When providing the query with the top command, the parse_one function previously raised an exception. I solved this by updating the tokens.py file and providing a new token type called top. This also solves the issue of converting the query with the top command to another query language, like PostgreSQL, which does not support the top command, and now, after this change, when using transpile, it adds the LIMIT keyword and provides the proper query.