Fix: use index2 instead of self._index in _parse_type index difference
I decided to slightly change this condition in order to use index2, which
is the index we arrive at right after parsing the type. The reason is that
between index2 and self._index, we may have somehow consumed more tokens
in the _parse_primary() call, and we don't want to take these into account.

I haven't thought of a code path where this happens, but I changed the
condition anyway because it seemed more robust this way.