Refactor: iteratively generate binary expressions (#3926)
Refactor the implementation for binary expressions in the generator
to use a partially iterative style, similar to `set_operations`.

The intention is to avoid unnecessary recursion while generating chains of N
binary operations, (e.g. a || b || c...), because if the chain gets too long,
the generator hits Python's recursion limit.

This isn't a completely iterative implementation, because we still call
the recursive `sql` method each time we encounter a different type of
Expression; but it works for these long chains.

To start with, only apply this optimisation for chains of the same operator.