fix: Add additional allowed tokens for parsing aggregate functions in DDL (#4727)
* Add test for failing AggregateFunction parsing

Added a test which catches cases where an AggregateFunction column
containing a parameterless `any` fails to parse.

* Allow parsing of additional token types for aggregate functions

Added ANY as an allowed TokenType when passing aggregate functions which
are a plain identifier and not a function call. This allows DDL for
dialects such as Clickhouse to correctly parse since the function names
used in the AggregateFunction column type may conflict with the `any`
keyword.