Support setting visibility on `ALTER COLUMN`.
MySQL 8.0 introduced the [ability to hide and unhide columns]. Columns
can be hidden in a `CREATE TABLE` statement (currently unimplemented in
`sqlglot`) and can be subsequently changed with an `ALTER TABLE` syntax:
`ALTER COLUMN x SET {VISIBLE|INVISIBLE}`.

This commit adds support for column hiding through the `ALTER TABLE`
syntax. However, it does so in a MySQL-specific way. Other RDBMSs have
similar support for column hiding, but implement it slightly
differently:

* Oracle uses `MODIFY x INVISIBLE`
* DB2 uses `ALTER x t IMPLICITLY HIDDEN`
* SQLServer uses `ALTER COLUMN x ADD HIDDEN`

As there is no other dialect in `sqlglot` that implements invisibility,
and because the parser currently maintains dialect-specific
implementations, I wasn't sure if the amount of effort to refactor the
parser was worthwhile for this change.

[ability to hide and unhide columns]: https://dev.mysql.com/doc/refman/8.4/en/alter-table.html