fix(snowflake): add support for FROM_HEX_BINARY
The [FROM_HEX_BINARY](https://docs.snowflake.com/en/sql-reference/functions/hex_decode_binary) frunction is equivalent to the `Unhex` function that is already supported by sqlglot. To support it, we just had to add the map in snowflake and add the transform in duckdb.

Partially solves #4852, we still need to add support for `FROM_HEX_STRING`.