fix!(tsql): support select...into #temp_table syntax
Fixes support for creating temporary tables as select statements in
tsql. The syntax is the same as selecting into a physical table, except
that tsql uses a `#` symbol to denote a temp table, and a `##` to denote
a global temp table.

This commit will slightly change the AST representation of the query,
such that the Table expression is marked as `temporary` rather than its
Identifier expression. This seems to be consistent with other dialects
that support `select...into` syntax, e.g. postgres