fix(tokenizer)!: add support for noop string escapes (#6526)
* feat!: add support for noop string escapes

This commit let dialects (like MySQL) allow escaping quotes that are not the current string delimiter, effectively treating the escape character as a no-op regarding the character's literal value but consuming the backslash. Previously, tokenize("""'\"a'""", "mysql") resulted in "a. With this change, it correctly results in "a.
Example:
In MySQL:
'"' -> " (Backslash is consumed)
'\"' -> " (Backslash is escaped, quote remains)

* fix!: add rust config for noop string escapes

* fix!: invert STRING_ESCAPES_NOOP to blacklist mode